+++ src/ApolloQueryObservable.ts
@@ -1,0 +1,3 @@
+import { RxObservableQuery } from 'apollo-client-rxjs';
+
+export class ApolloQueryObservable<T> extends RxObservableQuery<T> {}