+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### vNEXT
 
 - Added support for **ApolloClient** `v0.5.0` ([PR #5](https://github.com/apollostack/angular1-apollo/pull/5))
+- Added `watchQuery` and `subscribe` methods thanks to `apollo-client-rxjs`
 
 ### v0.0.1